Refactor connection options
