+++ src/cow-react/modules/wallet/api/pure/ConnectWalletOption/index.tsx
@@ -69,6 +69,7 @@ export const HeaderText = styled.div`
   /* color: ${(props) => (props.color === 'blue' ? ({ theme }) => theme.primary1 : ({ theme }) => theme.text1)}; */
   font-size: 1rem;
   font-weight: 500;
+  color: ${({ theme }) => theme.text1};
 `
 
 const SubHeader = styled.div`
