+++ src/cow-react/modules/wallet/web3-react/containers/ConnectWalletOptions/FortmaticOption.tsx
@@ -1,13 +1,28 @@
 import { ConnectionType } from '@cow/modules/wallet'
 
-import { fortmaticConnection } from '@cow/modules/wallet/web3-react/utils/connection'
+import Fortmatic from 'fortmatic'
 import { getConnectionName } from '@cow/modules/wallet/api/utils/connection'
 
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 import { ConnectWalletOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption'
 
 import { formaticOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption/ConnectWalletOptions'
-import { TryActivation } from '..'
+
+import { initializeConnector } from '@web3-react/core'
+import { EIP1193 } from '@web3-react/eip1193'
+
+import { TryActivation } from '.'
+import { Connection } from '../../utils/connection'
+
+
+const [web3Fortmatic, web3FortmaticHooks] = initializeConnector<EIP1193>(
+  (actions) => new EIP1193({ actions, provider: new Fortmatic(process.env.REACT_APP_FORTMATIC_KEY).getProvider() })
+)
+export const fortmaticConnection: Connection = {
+  connector: web3Fortmatic,
+  hooks: web3FortmaticHooks,
+  type: ConnectionType.FORTMATIC,
+}
 
 export function FortmaticOption({ tryActivation }: { tryActivation: TryActivation }) {
   const isActive = useIsActiveWallet(fortmaticConnection)
@@ -21,3 +36,4 @@ export function FortmaticOption({ tryActivation }: { tryActivation: TryActivatio
     />
   )
 }
+
