+++ src/cow-react/modules/wallet/web3-react/containers/ConnectWalletOptions/SafeOption.tsx
@@ -1,0 +1,15 @@
+
+import { initializeConnector } from '@web3-react/core'
+import { GnosisSafe } from '@web3-react/gnosis-safe'
+
+
+import { ConnectionType } from '@cow/modules/wallet'
+import { Connection } from '../../utils/connection'
+
+
+const [web3GnosisSafe, web3GnosisSafeHooks] = initializeConnector<GnosisSafe>((actions) => new GnosisSafe({ actions }))
+export const gnosisSafeConnection: Connection = {
+  connector: web3GnosisSafe,
+  hooks: web3GnosisSafeHooks,
+  type: ConnectionType.GNOSIS_SAFE,
+}
