+++ src/cow-react/modules/wallet/web3-react/containers/ConnectWalletOptions/WalletConnectOption.tsx
@@ -1,10 +1,36 @@
 import { ConnectionType } from '@cow/modules/wallet'
-import { walletConnectConnection } from '@cow/modules/wallet/web3-react/utils/connection'
 import { getConnectionName } from '@cow/modules/wallet/api/utils/connection'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 import { ConnectWalletOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption'
 import { walletConnectOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption/ConnectWalletOptions'
-import { TryActivation } from '..'
+import { TryActivation, onError } from '.'
+
+
+import { initializeConnector } from '@web3-react/core'
+import { WalletConnect } from '@web3-react/walletconnect'
+
+
+import { RPC_URLS } from 'constants/networks'
+import { Connection } from '../../utils/connection'
+
+
+const [web3WalletConnect, web3WalletConnectHooks] = initializeConnector<WalletConnect>(
+  (actions) =>
+    new WalletConnect({
+      actions,
+      options: {
+        rpc: RPC_URLS,
+        qrcode: true,
+      },
+      onError,
+    })
+)
+export const walletConnectConnection: Connection = {
+  connector: web3WalletConnect,
+  hooks: web3WalletConnectHooks,
+  type: ConnectionType.WALLET_CONNECT,
+}
+
 
 export function WalletConnectOption({ tryActivation }: { tryActivation: TryActivation }) {
   const isActive = useIsActiveWallet(walletConnectConnection)
