+++ src/hooks/useEagerlyConnect.ts
@@ -1,5 +1,6 @@
 import { Connector } from '@web3-react/types'
-import { gnosisSafeConnection, networkConnection } from '@cow/modules/wallet/web3-react/utils/connection'
+import { networkConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/NetworkOption'
+import { gnosisSafeConnection } from '@cow/modules/wallet/web3-react/containers/ConnectWalletOptions/SafeOption'
 import { getConnection } from '@cow/modules/wallet/web3-react/utils/connection'
 import { useEffect } from 'react'
 import { BACKFILLABLE_WALLETS } from 'state/connection/constants'
